backend-drf/requirements.txt
# Django and core framework
Django==4.1.13
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0

# Authentication
dj-rest-auth==2.2.6
django-allauth==0.54.0
requests-oauthlib==1.3.1

# Environment variables (for loading .env)
python-dotenv==1.0.1

# ASGI support (auto-installed with Django/channels, pinned for stability)
asgiref==3.6.0

# requirements.txt
Django==5.0.6                          
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1   # JWT token implementation for DRF
dj-rest-auth[with_social]==6.0.0       # REST API authentication utils


django-environ